Update by frame flag, not WAIT_FOR_VBLANK
Waiting for vblank via PPUSTATUS turns out to be a bad idea, since
reading PPUSTATUS unsets the vblank bit7 flag, which is used to trigger
the NMI launch.  (oops!)

We now use a zeropage register `frame` to track whether the frame is
done and to start the new update.  Much smoother!

Also some minor changes like an OAMDMA macro, as well as comments
explaining what is going on.