Animated bounce
Uses lookuptable to emulate a parabolic bounce.
Includes a python script to generate the table.

Much improvement here, but it looks better than I expected.
(There is some kind of frame-skipping effect as it hits the ground, I
dont know if that's me or just how this sort of thing looks when it's
fast.)