Formally convert pos/vel/acc to 2-byte
Not sure if this is actually more readable, but I guess it's a bit more
correct to define actual 16-bit regions for the y values used in
subpixel dynamics.

I may toggle this further (e.g. acc_y may *only* be subpixel) or may add
an alias for the high-byte for pos_y.  But despite the awkward +1 ness
of it, I think it's the right direction.