Jump physics to subroutine in jump.s
The jump solver is moved to a separate routine in a separate file.
Continuing the quest to clean up main.s

Downside is of course the jsr/rts but that's probably not much of an
issue.  Somewhat higher cost is the messier entanglement of memory with
import/export statements.  Though IMO that's just assembly for you.

We could further clean this up, although I don't know at what point we
are consuming precious cycles and bytes for the sake of clean code.