New platforms, 0.8 acc_y resolution
Platforms are now defined in this static lists of values:
- start x0
- end x1
- height y0
The static #160 check now looks for

This is, as one would expect, a bit weird due to scrolling, and of
course the graphics don't match, but it is more or less functional.

Acceleration was also dropped down from 8.8 to 0.8 resolution, since it
will never get anywhere near 1 pixel/frame.  (At least, I don't think
so!)