Platforms as ground?
Although this patch seems worse, I think it represents platform
progress.

We now do "delta" checks with respect to crossing platforms, rather than
just whether we are below the platform ("ground" checks).

The good news is that it seems to work, at least for the bottom
platform.

The bad news is that the downward acceleration is constant, and creates
a "delta" across the platform.  It seems to be causing a "stop" at the
bottom, then passing through on the next iteration.

I'm unsure if this should be solved by halting g to zero, or if this is
an off-by-one that can allow the acceleration to persist.  I'm leaning
towards the latter.

NOTE: Why did it work before?  Because we were also hard-setting y to
the ground value.  So we were constantly accelerating downward, then
just readjusting back up to ground.

Obviously this was never seen, because all of these calculation happen
inbetween the frame drawing.  The magic of video games!