No jump after running off cliff
The latch is now set whenever v < 0.  This includes any fall, which
advantageously captures the "run off a cliff" issue.

I also fixed an odd bug where x=255 would not be tested, since we were
always doing pos_x >= x1(p).  That is, x1(p) was being excluded from the
platform.  Maybe not a big deal in general, but in the current
(admittedly simple) form, it's otherwise impossible to do a full-screen
platform from 0 to 255.

Anyway.  Improvement!